backend_minimal/requirements.txt
fastapi
uvicorn
requests
opencv-python
numpy
ultralytics
pyyaml
